chore(OR-136): Backup tool in Rust - move ic.json5 to bin

+++ rs/backup/src/config.rs
@@ -12,7 +12,7 @@ pub struct SubnetConfig {
         serialize_with = "crate::util::replica_to_string"
     )]
     pub initial_replica_version: ReplicaVersion,
-    pub nodes_syncing: u32,
+    pub nodes_syncing: usize,
     pub sync_period_secs: u64,
     pub replay_period_secs: u64,
 }
